new classifier ic with lowest 33%

+++ SimCLR/permclr.py
@@ -194,7 +194,7 @@ class PermCLR(object):
 				#Count how many of the indices before 10 belongs to each of the three classes
 				indices = indices[:train_batch_size]
 				for j in range(num_classes):
-					new_logits[i*(num_classes) + j ] = torch.sum((train_batch_size*j<=indices) * (indices<train_batch_size*(j+1)))/train_batch_size
+					new_logits[i*(num_classes) + j ] = torch.sum((train_batch_size*j<=indices) * (indices<train_batch_size*(j+1))).float()/train_batch_size
